[COZY-481] rename : Converter를 /util 에서 /converter로 이동
